requirements.txt
boto3==1.17.45
botocore==1.20.45
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-auth==1.28.0
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.1
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
packaging==20.9
pip==21.0.1
protobuf==3.15.7
pyasn1-modules==0.2.8
pyasn1==0.4.8
pygsheets==2.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7.2
s3transfer==0.3.6
schedule==1.0.0
setuptools==54.2.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.4
python-dotenv==0.17.0
json-logging==1.2.11
